requirements.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.7.0 ; python_version >= "3.11" and python_version < "4"
cachetools==5.3.3 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
cloudevents==1.10.1 ; python_version >= "3.11" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
deprecated==1.2.14 ; python_version >= "3.11" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4"
flask==3.0.2 ; python_version >= "3.11" and python_version < "4"
functions-framework==3.5.0 ; python_version >= "3.11" and python_version < "4"
google-api-core[grpc]==2.17.1 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.28.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-secret-manager==2.18.2 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-trace==1.13.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.62.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.13.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.62.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.62.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4" and platform_system != "Windows"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.4 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==6.11.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4"
jinja2==3.1.3 ; python_version >= "3.11" and python_version < "4"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4"
opentelemetry-api==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-exporter-gcp-trace==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-common==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-grpc==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-instrumentation-httpx==0.44b0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-instrumentation-wsgi==0.44b0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-instrumentation==0.44b0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-propagator-gcp==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-proto==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-resourcedetector-gcp==1.6.0a0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-sdk==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-semantic-conventions==0.44b0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-util-http==0.44b0 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4"
proto-plus==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.16.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
setuptools==69.1.1 ; python_version >= "3.11" and python_version < "4.0"
slack-sdk==3.27.1 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
structlog-gcp==0.2.4 ; python_version >= "3.11" and python_version < "4.0"
structlog==24.1.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
watchdog==4.0.0 ; python_version >= "3.11" and python_version < "4"
werkzeug==3.0.1 ; python_version >= "3.11" and python_version < "4"
wrapt==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.11" and python_version < "4.0"
